requirements.txt
huggingface-hub
pymongo
python-dotenv
